+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.0] - 2025-12-04
+
+### Added
+
+- Ability to tell openmock to use a real server.
+- Extra `openmock[web]`, which enables `openmock` cli command which runs a web UI for a web console for openmock.
+
+### Fixed
+
+- Various issues found when running tests against OpenSearch.
+
+
 ## [3.1.5] - 2025-12-04
 
 ### Added